HSEARCH-5033 Better tests and documentation of limitations around nested fields for highlighting
